Screen-to-text fixes in VM test script
ImageMagick (again) updated the internal algorithm used to compute image
signatures. That broke the OCR as the hashes were used to identify
individual characters.

This commit introduced a new mechanism that relies on determining color
of individual pixels when converting to text. Hopefully this mechanism
will be more stable :-).

The OCR patterns are now in separate file to simplify updating (or adding
new characters). Also a new switch --train-ocr was added for such task.