test-in-vm: workaround for rare vterm matching
It seems that under very specific conditions we might be in situation
when the tail of vterm against which we are checking is actually shorter
than the new screen. This is probably error on the caller side that has
not reset the buffer but untils this is investigated more, we keep this
extra catch block.